+++ AppKit/NSColor.subproj/NSColor.m
@@ -496,31 +496,31 @@ static void releasePatternInfo(void *info){
    NSInvalidAbstractInvocation();
 }
 
--(void)getWhite:(float *)white alpha:(float *)alpha {
+-(void)getWhite:(CGFloat *)white alpha:(CGFloat *)alpha {
     NSInvalidAbstractInvocation();
 }
 
--(void)getRed:(float *)red green:(float *)green blue:(float *)blue alpha:(float *)alpha {
+-(void)getRed:(CGFloat *)red green:(CGFloat *)green blue:(CGFloat *)blue alpha:(CGFloat *)alpha {
    NSInvalidAbstractInvocation();
 }
 
--(void)getHue:(float *)hue saturation:(float *)saturation brightness:(float *)brightness alpha:(float *)alpha {
+-(void)getHue:(CGFloat *)hue saturation:(CGFloat *)saturation brightness:(CGFloat *)brightness alpha:(CGFloat *)alpha {
    NSInvalidAbstractInvocation();
 }
 
--(void)getCyan:(float *)cyan magenta:(float *)magenta yellow:(float *)yellow black:(float *)black alpha:(float *)alpha {
+-(void)getCyan:(CGFloat *)cyan magenta:(CGFloat *)magenta yellow:(CGFloat *)yellow black:(CGFloat *)black alpha:(CGFloat *)alpha {
     NSInvalidAbstractInvocation();
 }
 
 -(CGFloat)whiteComponent {
-    float white;
+    CGFloat white;
 
     [self getWhite:&white alpha:NULL];
     return white;
 }
 
 -(CGFloat)redComponent {
-   float red;
+   CGFloat red;
 
    [self getRed:&red green:NULL blue:NULL alpha:NULL];
 
@@ -528,7 +528,7 @@ static void releasePatternInfo(void *info){
 }
 
 -(CGFloat)greenComponent {
-   float green;
+   CGFloat green;
 
    [self getRed:NULL green:&green blue:NULL alpha:NULL];
 
@@ -536,7 +536,7 @@ static void releasePatternInfo(void *info){
 }
 
 -(CGFloat)blueComponent {
-   float blue;
+   CGFloat blue;
 
    [self getRed:NULL green:NULL blue:&blue alpha:NULL];
 
@@ -544,7 +544,7 @@ static void releasePatternInfo(void *info){
 }
 
 -(CGFloat)hueComponent {
-   float hue;
+   CGFloat hue;
 
    [self getHue:&hue saturation:NULL brightness:NULL alpha:NULL];
 
@@ -552,7 +552,7 @@ static void releasePatternInfo(void *info){
 }
 
 -(CGFloat)saturationComponent {
-   float saturation;
+   CGFloat saturation;
 
    [self getHue:NULL saturation:&saturation brightness:NULL alpha:NULL];
 
@@ -560,7 +560,7 @@ static void releasePatternInfo(void *info){
 }
 
 -(CGFloat)brightnessComponent {
-   float brightness;
+   CGFloat brightness;
 
    [self getHue:NULL saturation:NULL brightness:&brightness alpha:NULL];
 
@@ -568,28 +568,28 @@ static void releasePatternInfo(void *info){
 }
 
 -(CGFloat)cyanComponent {
-    float cyan;
+    CGFloat cyan;
 
     [self getCyan:&cyan magenta:NULL yellow:NULL black:NULL alpha:NULL];
     return cyan;
 }
 
 -(CGFloat)magentaComponent {
-    float magenta;
+    CGFloat magenta;
 
     [self getCyan:NULL magenta:&magenta yellow:NULL black:NULL alpha:NULL];
     return magenta;
 }
 
 -(CGFloat)yellowComponent {
-    float yellow;
+    CGFloat yellow;
 
     [self getCyan:NULL magenta:NULL yellow:&yellow black:NULL alpha:NULL];
     return yellow;
 }
 
 -(CGFloat)blackComponent {
-    float black;
+    CGFloat black;
 
     [self getCyan:NULL magenta:NULL yellow:NULL black:&black alpha:NULL];
     return black;
@@ -624,9 +624,9 @@ static void releasePatternInfo(void *info){
    if(primary==nil || secondary==nil)
     return nil;
    else {
-    float pr,pg,pb,pa;
-    float sr,sg,sb,sa;
-    float rr,rg,rb,ra;
+    CGFloat pr,pg,pb,pa;
+    CGFloat sr,sg,sb,sa;
+    CGFloat rr,rg,rb,ra;
 
     [primary getRed:&pr green:&pg blue:&pb alpha:&pa];
     [secondary getRed:&sr green:&sg blue:&sb alpha:&sa];
