+++ AppKit/NSGradient.m
@@ -15,7 +15,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 
 @implementation NSGradient
 
-static void evaluate(void *info,float const *input,float *output) {
+static void evaluate(void *info,CGFloat const *input,CGFloat *output) {
    NSGradient *self=(NSGradient *)info;
    CGFloat   **components=self->_components;
    CGFloat    *locations=self->_locations;
@@ -243,8 +243,8 @@ static void evaluate(void *info,float const *input,float *output) {
 }
 
 -(NSColor *)interpolatedColorAtLocation:(CGFloat)location {
-   float input[1]={location};
-   float output[4];
+   CGFloat input[1]={location};
+   CGFloat output[4];
    
    evaluate(self,input,output);
