+++ CoreGraphics/CGContext.m
@@ -220,7 +220,7 @@ void CGContextSetFillColorSpace(CGContextRef context,CGColorSpaceRef colorSpace)
    O2ContextSetFillColorSpace(context,colorSpace);
 }
 
-void CGContextSetStrokeColor(CGContextRef context,const float *components) {
+void CGContextSetStrokeColor(CGContextRef context,const CGFloat *components) {
    O2ContextSetStrokeColor(context,components);
 }
 
@@ -240,7 +240,7 @@ void CGContextSetCMYKStrokeColor(CGContextRef context,float c,float m,float y,fl
    O2ContextSetCMYKStrokeColor(context,c,m,y,k,alpha);
 }
 
-void CGContextSetFillColor(CGContextRef context,const float *components) {
+void CGContextSetFillColor(CGContextRef context,const CGFloat *components) {
    O2ContextSetFillColor(context,components);
 }
